requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
djangorestframework==3.11.0
Faker==4.0.3
gunicorn==20.0.4
psycopg2==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.0.1
